refactor #62: location 관련 fetcher 함수들 파일 분리
